Remove wrappers and driver.
