+++ metal.go
@@ -53,8 +53,8 @@ type Client interface {
 	VPN() vpn.ClientService
 }
 
-// Driver holds the client connection to the metal api
-type Driver struct {
+// driver holds the client connection to the metal api
+type driver struct {
 	c *client.MetalAPI
 
 	bearer       string
@@ -63,30 +63,30 @@ type Driver struct {
 }
 
 // Option for config of Driver
-type option func(driver *Driver)
+type option func(driver *driver)
 
 // AuthType sets the authType for HMAC-Auth
 func AuthType(authType string) option {
-	return func(driver *Driver) {
+	return func(driver *driver) {
 		driver.hmacAuthType = authType
 	}
 }
 
 // NewDriver Create a new Driver for Metal to given url. Either bearer OR hmacKey must be set.
 // The returned *Driver will be deprecated at some point in time, please migrate to use the Client interface instead.
-func NewDriver(baseURL, bearer, hmacKey string, options ...option) (Client, *Driver, error) {
+func NewDriver(baseURL, bearer, hmacKey string, options ...option) (Client, error) {
 	parsedURL, err := url.Parse(baseURL)
 	if err != nil {
-		return nil, nil, err
+		return nil, err
 	}
 	if parsedURL.Host == "" {
-		return nil, nil, fmt.Errorf("invalid url:%s, must be in the form scheme://host[:port]/basepath", baseURL)
+		return nil, fmt.Errorf("invalid url:%s, must be in the form scheme://host[:port]/basepath", baseURL)
 	}
 
 	transport := httptransport.New(parsedURL.Host, parsedURL.Path, []string{parsedURL.Scheme})
 	c := client.New(transport, strfmt.Default)
 
-	driver := &Driver{
+	driver := &driver{
 		c:            c,
 		bearer:       bearer,
 		hmacAuthType: defaultHMACAuthType,
@@ -105,10 +105,10 @@ func NewDriver(baseURL, bearer, hmacKey string, options ...option) (Client, *Dri
 
 	// TODO: remove *Driver return at some point in the future in order to get rid off the handwritten wrappers
 	// see: https://github.com/metal-stack/metal-go/issues/33
-	return driver, driver, nil
+	return driver, nil
 }
 
-func (d *Driver) auther(rq runtime.ClientRequest, rg strfmt.Registry) error {
+func (d *driver) auther(rq runtime.ClientRequest, rg strfmt.Registry) error {
 	if d.hmac != nil {
 		d.hmac.AddAuthToClientRequest(rq, time.Now())
 	} else if d.bearer != "" {
@@ -117,78 +117,54 @@ func (d *Driver) auther(rq runtime.ClientRequest, rg strfmt.Registry) error {
 	return nil
 }
 
-func StrDeref(s *string) string {
-	if s == nil {
-		var res string
-		return res
-	}
-	return *s
-}
-
-func BoolDeref(b *bool) bool {
-	if b == nil {
-		var res bool
-		return res
-	}
-	return *b
-}
-
-func Int64Deref(i *int64) int64 {
-	if i == nil {
-		var res int64
-		return res
-	}
-	return *i
-}
-
-func (d *Driver) Filesystemlayout() filesystemlayout.ClientService {
+func (d *driver) Filesystemlayout() filesystemlayout.ClientService {
 	return d.c.Filesystemlayout
 }
-func (d *Driver) Firewall() firewall.ClientService {
+func (d *driver) Firewall() firewall.ClientService {
 	return d.c.Firewall
 }
-func (d *Driver) Firmware() firmware.ClientService {
+func (d *driver) Firmware() firmware.ClientService {
 	return d.c.Firmware
 }
-func (d *Driver) Health() health.ClientService {
+func (d *driver) Health() health.ClientService {
 	return d.c.Health
 }
-func (d *Driver) Image() image.ClientService {
+func (d *driver) Image() image.ClientService {
 	return d.c.Image
 }
-func (d *Driver) IP() ip.ClientService {
+func (d *driver) IP() ip.ClientService {
 	return d.c.IP
 }
-func (d *Driver) Machine() machine.ClientService {
+func (d *driver) Machine() machine.ClientService {
 	return d.c.Machine
 }
-func (d *Driver) Network() network.ClientService {
+func (d *driver) Network() network.ClientService {
 	return d.c.Network
 }
-func (d *Driver) Partition() partition.ClientService {
+func (d *driver) Partition() partition.ClientService {
 	return d.c.Partition
 }
-func (d *Driver) Project() project.ClientService {
+func (d *driver) Project() project.ClientService {
 	return d.c.Project
 }
-func (d *Driver) Size() size.ClientService {
+func (d *driver) Size() size.ClientService {
 	return d.c.Size
 }
-func (d *Driver) Sizeimageconstraint() sizeimageconstraint.ClientService {
+func (d *driver) Sizeimageconstraint() sizeimageconstraint.ClientService {
 	return d.c.Sizeimageconstraint
 }
-func (d *Driver) SwitchOperations() switch_operations.ClientService {
+func (d *driver) SwitchOperations() switch_operations.ClientService {
 	return d.c.SwitchOperations
 }
-func (d *Driver) Tenant() tenant.ClientService {
+func (d *driver) Tenant() tenant.ClientService {
 	return d.c.Tenant
 }
-func (d *Driver) User() user.ClientService {
+func (d *driver) User() user.ClientService {
 	return d.c.User
 }
-func (d *Driver) Version() version.ClientService {
+func (d *driver) Version() version.ClientService {
 	return d.c.Version
 }
-func (d *Driver) VPN() vpn.ClientService {
+func (d *driver) VPN() vpn.ClientService {
 	return d.c.Vpn
 }
